+++ packages/build-tools/src/utils/buildEnv.ts
